requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
uvicorn==0.34.3
aiohttp==3.12.12
azure-core==1.34.0